# Optimise alignment/part2 optimum tracker policy