--- a/include/seqan3/alignment/pairwise/detail/pairwise_alignment_algorithm.hpp
+++ b/include/seqan3/alignment/pairwise/detail/pairwise_alignment_algorithm.hpp
@@ -165,54 +165,137 @@ protected:
         // Initialisation phase: allocate memory and initialise first column.
         // ---------------------------------------------------------------------
 
+        this->reset_optimum(); // Reset the tracker for the new alignment computation.
+
         thread_local score_matrix_single_column<int32_t> local_matrix{};
         local_matrix.reset_matrix(std::forward<sequence1_t>(sequence1), std::forward<sequence2_t>(sequence2));
         auto matrix_iter = local_matrix.begin();
-        auto alignment_column = *matrix_iter;
 
-        // Initialise the first column.
-        *alignment_column.begin() = this->initialise_origin_cell();
-        for (auto cell : alignment_column | seqan3::views::drop(1))
-            cell = this->initialise_first_column_cell(cell);
+        initialise_column(*matrix_iter, sequence2);
 
         // ---------------------------------------------------------------------
-        // Recursion phase: compute column-wise the alignment matrix.
+        // Iteration phase: compute column-wise the alignment matrix.
         // ---------------------------------------------------------------------
 
-        // Compute remaining matrix
-        ++matrix_iter; // Move to next column.
-        for (auto sequence1_iter = sequence1.begin();
-             sequence1_iter != sequence1.end();
-             ++sequence1_iter, ++matrix_iter)
+        for (auto sequence1_value : sequence1)
         {
-            alignment_column = *matrix_iter;
-            auto alignment_column_iter = alignment_column.begin();
-
-            // Initialise first cell of current column.
-            auto cell = *alignment_column_iter;
-            typename traits_type::score_type diagonal = cell.optimal_score();
-            *alignment_column_iter = this->initialise_first_row_cell(cell);
-
-            ++alignment_column_iter;  // Move to next cell in column.
-            for (auto sequence2_iter = sequence2.begin();
-                 sequence2_iter != sequence2.end();
-                 ++sequence2_iter, ++alignment_column_iter)
-            {
-                auto cell = *alignment_column_iter;
-                typename traits_type::score_type next_diagonal = cell.optimal_score();
-                *alignment_column_iter = this->compute_inner_cell(diagonal,
-                                                                  cell,
-                                                                  m_scoring_scheme.score(*sequence1_iter,
-                                                                                         *sequence2_iter));
-                diagonal = next_diagonal;
-            }
+            ++matrix_iter;
+            compute_column(*matrix_iter, sequence1_value, sequence2);
         }
 
         // ---------------------------------------------------------------------
-        // Wrap up phase: track score of last column
+        // Final phase: track score of last column
         // ---------------------------------------------------------------------
 
-        return alignment_column[std::ranges::distance(sequence2)].optimal_score();
+        auto alignment_column = *matrix_iter;
+        auto column_iterator = alignment_column.begin();
+        this->track_last_column_cell(*column_iterator);
+
+        for ([[maybe_unused]] auto && unused : sequence2)
+            this->track_last_column_cell(*++column_iterator);
+
+        this->track_final_cell(*column_iterator);
+
+        return this->tracked_optimum();
+    }
+
+    /*!\brief Initialise the first column of the alignment matrix.
+     * \tparam alignment_matrix_column_t The type of the alignment matrix; must model std::ranges::input_range.
+     * \tparam sequence2_t The type of the second sequence; must model std::ranges::input_range.
+     *
+     * \param[in] first_column The first column of the alignment matrix to initialise.
+     * \param[in] sequence2 The second sequence used to determine the size of the column.
+     *
+     * \details
+     *
+     * The first column of the alignment matrix does not require any character comparisons of the sequences that
+     * shall be aligned. The second sequence is thus only needed to determine the size of the column.
+     * The computation of the column is split into three phases: the initialisation phase, the iteration phase, and
+     * the final phase. In the initialisation phase the first cell of the column is computed and in the iteration
+     * phase all remaining cells are computed. In the final phase the last cell is possibly evaluated for a new
+     * alignment optimum.
+     */
+    template <std::ranges::input_range alignment_matrix_column_t, std::ranges::input_range sequence2_t>
+    void initialise_column(alignment_matrix_column_t && first_column, sequence2_t && sequence2)
+    {
+        // ---------------------------------------------------------------------
+        // Initial phase: prepare column and initialise first cell
+        // ---------------------------------------------------------------------
+
+        auto first_column_iter = first_column.begin();
+        *first_column_iter = this->track_cell(this->initialise_origin_cell());
+
+        // ---------------------------------------------------------------------
+        // Iteration phase: iterate over column and compute each cell
+        // ---------------------------------------------------------------------
+
+        for ([[maybe_unused]] auto && unused : sequence2)
+        {
+            ++first_column_iter;
+            *first_column_iter = this->track_cell(this->initialise_first_column_cell(*first_column_iter));
+        }
+
+        // ---------------------------------------------------------------------
+        // Final phase: track last cell of initial column
+        // ---------------------------------------------------------------------
+
+        this->track_last_row_cell(*first_column_iter);
+    }
+
+    /*!\brief Initialise any column of the alignment matrix except the first one.
+     * \tparam alignment_matrix_column_t The type of the alignment matrix; must model std::ranges::input_range.
+     * \tparam sequence1_value_t The value type of sequence1; must model seqan3::semialphabet.
+     * \tparam sequence2_t The type of the second sequence; must model std::ranges::input_range.
+     *
+     * \param[in] alignment_column The column of the alignment matrix to compute.
+     * \param[in] sequence1_value The current symbol of sequence1.
+     * \param[in] sequence2 The second sequence to align against `sequence1_value`.
+     *
+     * \details
+     *
+     * Computes the alignment for the given alignment matrix column. The function splits the computation of the column
+     * into three phases: the initialisation phase, the iteration phase, and the final phase. In the initialisation
+     * phase the first cell of the column is computed and in the iteration phase all remaining cells are computed.
+     * In the final phase the last cell is possibly evaluated for a new alignment optimum.
+     */
+    template <std::ranges::input_range alignment_matrix_column_t,
+              semialphabet sequence1_value_t,
+              std::ranges::input_range sequence2_t>
+    void compute_column(alignment_matrix_column_t && alignment_column,
+                        sequence1_value_t const & sequence1_value,
+                        sequence2_t && sequence2)
+    {
+        using score_type = typename traits_type::score_type;
+
+        // ---------------------------------------------------------------------
+        // Initial phase: prepare column and initialise first cell
+        // ---------------------------------------------------------------------
+
+        auto alignment_column_iter = alignment_column.begin();
+
+        auto cell = *alignment_column_iter;
+        score_type diagonal = cell.optimal_score();
+        *alignment_column_iter = this->track_cell(this->initialise_first_row_cell(cell));
+
+        // ---------------------------------------------------------------------
+        // Iteration phase: iterate over column and compute each cell
+        // ---------------------------------------------------------------------
+
+        for (auto && sequence2_value : sequence2)
+        {
+            ++alignment_column_iter;
+            auto cell = *alignment_column_iter;
+            score_type next_diagonal = cell.optimal_score();
+            *alignment_column_iter = this->track_cell(
+                this->compute_inner_cell(diagonal, cell, m_scoring_scheme.score(sequence1_value, sequence2_value)));
+            diagonal = next_diagonal;
+        }
+
+        // ---------------------------------------------------------------------
+        // Final phase: track last cell
+        // ---------------------------------------------------------------------
+
+        this->track_last_row_cell(*alignment_column_iter);
     }
 };