In the new alignment implementation we let the alignment algorithm decide when to check for the new alignment optimum. This removes the dependency between the recursion policy and the tracker policy as we used it before. This in turn allows us to remove the CRTP design from the alignment algorithm. The performance is still very competitive as the microbenchmark shows below. My macrobenchmarks on this design confirmed this as well.

This includes only the pre-work and does only work for the global alignment so far. My goal is to get as quickly as possible all minimal requirements into the new alignment design to provide a working vectorised version which has the optimal runtime of SeqAn2. For this I need to provide a special tracker that overloads the current one in order to check the end points for sequences that have not the same size in the batch. This is a general problem with our inter-sequence vectorisation design and my initial benchmarks show for the new version a 1.5x speedup over the implementation in SeqAn2.

Old microbenchmark of global affine alignment without vectorisation:
```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
seqan3_affine_dna4                      912108 ns       901837 ns          724 CUPS=278.322M/s cells=251.001k
seqan2_affine_dna4                      507343 ns       501442 ns         1293 CUPS=500.558M/s cells=251.001k

seqan3_affine_dna4_collection          3780352 ns      3737470 ns          185 CUPS=272.939M/s cells=1020.1k
seqan2_affine_dna4_collection          2544802 ns      2516513 ns          273 CUPS=405.363M/s cells=1020.1k
```

New microbenchmark of global affine alignment without vectorisation:

```
----------------------------------------------------------------------------------------------
Benchmark                                    Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------------------------
seqan3_affine_dna4                      508742 ns       501681 ns         1315 CUPS=500.32M/s cells=251.001k
seqan2_affine_dna4                      506287 ns       500708 ns         1280 CUPS=501.292M/s cells=251.001k

seqan3_affine_dna4_collection          2098579 ns      2070642 ns          332 CUPS=492.649M/s cells=1020.1k
seqan2_affine_dna4_collection          2559029 ns      2508209 ns          273 CUPS=406.705M/s cells=1020.1k
```